pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cerberus-obfuscator"
version = "0.1.0"
description = "Advanced multi-layer Python obfuscator with runtime protections."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "gvoze32"}]
dependencies = [
  "pycryptodome>=3.23.0",
  "requests>=2.32.0",
  "psutil>=5.9.0",
  "nuitka>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.5",
  "mypy>=1.11",
]

[project.scripts]
cerberus = "cerberus.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cerberus*"]


